requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
colorama==0.4.4
fastapi==0.75.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
